requirements.txt
email_validator==2.2.0
fastapi==0.115.5
httpx==0.27.2
pipdeptree==2.23.4
pydantic-settings==2.6.1
python-multipart==0.0.17
PyYAML==6.0.2
resend==2.4.0
uvicorn==0.32.0
cerberus==1.3.5
jsonschema==4.23.0
